// capstone-frontend/src/app/models/person.ts
export class Person {
    id?: string; // Optional, as it may be generated by the backend
    name: string;
    role: string;
    email: string;
    password: string;
    phone: string;
    address: string;
    gender: string;
    age: number;
    emergencyContact: string;
  
    constructor(
      name: string,
      role: string,
      email: string,
      password: string,
      phone: string,
      address: string,
      gender: string,
      age: number,
      emergencyContact: string
    ) {
      this.name = name;
      this.role = role;
      this.email = email;
      this.password = password;
      this.phone = phone;
      this.address = address;
      this.gender = gender;
      this.age = age;
      this.emergencyContact = emergencyContact;
    }
  }
  